refactor: rename navigation.js to navOperations.js
